match decrypt_magic_codes

+++ src/objects.h
@@ -517,7 +517,7 @@ s32 obj_animate(Object *obj);
 //Non Matching
 void calc_dynamic_lighting_for_object_1(Object *, ObjectModel *, s16, Object *, f32, f32);
 void calc_dynamic_lighting_for_object_2(Object *, ObjectModel *, s16, f32);
-void decrypt_magic_codes(s32 *data, s32 length);
+void decrypt_magic_codes(u8 *data, s32 length);
 s32 func_80014814(s32 *);
 Object *spawn_object(LevelObjectEntryCommon *entry, s32);
 s32 func_8001F460(Object*, s32, Object*);
